JIRA board: Link card titles to issue: Fix for updated site
With the titles wrapped in links, the page was showing an error on hover of a title. I think this is because it's expecting elements to be in particular states (for replacement) when it's trying to add in the issue title edit button. I've fixed it by instead wrapping an element that's a few nodes up the stack